// shared/sdk/src/fields/field.enum.ts
export enum FieldFamily {
  Normal = 'normal',
  Start = 'start',
  Action = 'action',
  Bank = 'bank',
  Bonus = 'bonus',
  Tax = 'tax',
  Jail = 'jail',
  Station = 'station',
  Factory = 'factory',
  StreetGrey = 'street-grey',
  StreetYellow = 'street-yellow',
  StreetOrange = 'street-orange',
  StreetGreen = 'street-green',
  StreetPurple = 'street-purple',
  StreetWhite = 'street-white',
  StreetBlue = 'street-blue',
  StreetRed = 'street-red',
}